SY-2965: Clean up 2
