+++ integration/test/console/log.py
@@ -7,32 +7,27 @@
 #  License, use of this software will be governed by the Apache License, Version 2.0,
 #  included in the file licenses/APL.txt.
 
-from typing import TYPE_CHECKING, Any, Dict, List, Optional, Union
+from typing import TYPE_CHECKING, List, Optional
 
 from playwright.sync_api import Page
 
-from ..console_page import ConsolePage
+from .console_page import ConsolePage
 
 if TYPE_CHECKING:
     from ..console import Console
 
 
-class Table(ConsolePage):
-    """Table page management interface"""
+class Log(ConsolePage):
+    """Log page management interface"""
 
     def __init__(self, page: Page, console: "Console") -> None:
         super().__init__(page, console)
-        self.page_type = "Table"
-        self.pluto_label = ".pluto-table"
+        self.page_type = "Log"
+        self.pluto_label = ".pluto-log"
 
-    def add_column(self) -> None:
+    def clear(self) -> None:
+        """Clear all log entries."""
         pass
 
-    def remove_column(self) -> None:
-        pass
-
-    def get_row_count(self) -> int:
-        return 0
-
-    def edit_cell(self, row: int, column: str, value: Any) -> None:
-        pass
+    def get_entries(self, level: Optional[str] = None) -> List[str]:
+        return []
