+++ integration/test/console/table.py
@@ -7,27 +7,32 @@
 #  License, use of this software will be governed by the Apache License, Version 2.0,
 #  included in the file licenses/APL.txt.
 
-from typing import TYPE_CHECKING, List, Optional
+from typing import TYPE_CHECKING, Any, Dict, List, Optional, Union
 
 from playwright.sync_api import Page
 
-from ..console_page import ConsolePage
+from .console_page import ConsolePage
 
 if TYPE_CHECKING:
     from ..console import Console
 
 
-class Log(ConsolePage):
-    """Log page management interface"""
+class Table(ConsolePage):
+    """Table page management interface"""
 
     def __init__(self, page: Page, console: "Console") -> None:
         super().__init__(page, console)
-        self.page_type = "Log"
-        self.pluto_label = ".pluto-log"
+        self.page_type = "Table"
+        self.pluto_label = ".pluto-table"
 
-    def clear(self) -> None:
-        """Clear all log entries."""
+    def add_column(self) -> None:
         pass
 
-    def get_entries(self, level: Optional[str] = None) -> List[str]:
-        return []
+    def remove_column(self) -> None:
+        pass
+
+    def get_row_count(self) -> int:
+        return 0
+
+    def edit_cell(self, row: int, column: str, value: Any) -> None:
+        pass
